feat(core): Updates CatalogItemDetailsFragmentTest to BookingItemDetailsFragmentTest.
